Remove current project version from the sidebar
Without the ability to switch versions, showing the latest version of a project brings minimal use.

It has a cost, though. Because we track the latest changes on the master branch, we might be showing in the documentation new features that have not yet been released, yet the version indicator in the corner implies they are, misleading the users.

Ultimately this problem is caused by npm’s package.json insistence on keeping the version field inside the configuration, where it usually has no meaning. After all, for all non-release commits the package.json version is lying.