Check buildForRoot config flag
This code will check if the buildForRoot flag is set to true. If so, it will also check if there is a baseURL provided. If so, the filed outputDir will be stripped of the baseURL. Otherwise, the outputDir stays the same. This will allow for the outputted "items" to be in the root directory.